requirements.txt
asgiref==3.7.2
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
chardet==3.0.4
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.7
cssselect==1.2.0
Django==4.2.9
django-admin-list-filter-dropdown==1.0.3
django-tinymce==3.6.1
django-treebeard==4.7.1
djangorestframework==3.15.1
filelock==3.13.1
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.1.5
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==5.1.0
packaging==23.2
parsel==1.8.1
Protego==0.3.0
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.3.0
python-dotenv==1.0.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
rfc3986==1.5.0
Scrapy==2.11.0
scrapy-djangoitem==1.1.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
tldextract==5.1.1
Twisted==22.10.0
typing_extensions==4.9.0
urllib3==2.1.0
w3lib==2.1.2
zope.interface==6.1
